added tools blueprint for api
