+++ src/lanscape/blueprints/api/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 api_bp = Blueprint('api', __name__)
 
-from . import port, scan
+from . import port, scan, tools
