+++ src/lanscape/blueprints/api/scan.py
@@ -56,16 +56,3 @@ def scan_subnet_async():
 def get_scan(scan_id):
     scan = SubnetScanner.get_scan(scan_id)
     return jsonify(scan)
-
-@api_bp.route('/api/scan/subnet/test')
-def test_subnet():
-    subnet = request.args.get('subnet')
-    if not subnet: return jsonify({'valid': False, 'msg': 'Subnet cannot be blank'})
-    try:
-        ips = parse_ip_input(subnet)
-        length = len(ips)
-        return jsonify({'valid': True, 'msg': f"{length} IP{'s' if length > 1 else ''}"})
-    except SubnetTooLargeError:
-        return jsonify({'valid': False, 'msg': 'subnet too large', 'error': traceback.format_exc()})
-    except:
-        return jsonify({'valid': False, 'msg': 'invalid subnet', 'error': traceback.format_exc()})
